pyproject.toml
[tool.poetry]
name = "lemonade-home"
version = "0.1.0"
description = ""
authors = ["moshe <moshe@mosh>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pyssandra = "^0.12.4"
python-dotenv = "^1.0.0"
pyspark = "^3.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"